backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
torchaudio==2.3.1
soundfile==0.12.1
audioread==3.0.1
ollama==0.1.6
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
numba==0.58.1
